refactor: Update AuthLogin service to use new error class
